fix and move test and add documentation
