+++ src/Features/SupportDataBinding/BrowserTest.php
@@ -2,9 +2,10 @@
 
 namespace Livewire\Features\SupportDataBinding;
 
-use Livewire\Component;
-use Livewire\Livewire;
 use Tests\BrowserTestCase;
+use Livewire\Livewire;
+use Livewire\Component;
+use Livewire\Attributes\Computed;
 
 class BrowserTest extends BrowserTestCase
 {
@@ -34,5 +35,60 @@ class BrowserTest extends BrowserTestCase
             ->assertSee('The data is in-sync...')
             ->assertDontSee('Unsaved changes...')
         ;
+    }
+
+    /** @test */
+    public function updates_dependent_select_options_correctly_when_wire_key_is_applied()
+    {
+        Livewire::visit(new class extends Component {
+            public $parent = 'foo';
+
+            public $child = 'bar';
+
+            protected $options = [
+                'foo' => [
+                    'bar',
+                ],
+                'baz' => [
+                    'qux',
+                ],
+            ];
+
+            #[Computed]
+            public function parentOptions(): array
+            {
+                return array_keys($this->options);
+            }
+
+            #[Computed]
+            public function childOptions(): array
+            {
+                return $this->options[$this->parent];
+            }
+
+            public function render(): string
+            {
+                return <<<'blade'
+                    <div>
+                        <select wire:model.live="parent" dusk="parent">
+                            @foreach($this->parentOptions as $value)
+                                <option value="{{ $value }}">{{ $value }}</option>
+                            @endforeach
+                        </select>
+
+                        <select wire:model="child" dusk="child" wire:key="{{ $parent }}">
+                            <option value>Select</option>
+                            @foreach($this->childOptions as $value)
+                                <option value="{{ $value }}">{{ $value }}</option>
+                            @endforeach
+                        </select>
+                    </div>
+                blade;
+            }
+        })
+            ->waitForLivewire()->select('@parent', 'baz')
+            ->assertSelected('@child', '')
+            ->waitForLivewire()->select('@parent', 'foo')
+            ->assertSelected('@child', 'bar');
     }
 }
